requirements.txt
# Telegram
python-telegram-bot[job-queue]==21.7

# LLM providers
anthropic>=0.40.0
openai>=1.50.0

# HTTP + Content extraction
httpx>=0.27.0
beautifulsoup4>=4.12.0
readability-lxml>=0.8.1
lxml>=5.0.0

# Database
aiosqlite>=0.20.0

# Utils
python-dotenv==1.0.1
